Resuse code for getting keys in serverless code.
Several of the serverless commands (client, integrate and
sequence) use nearly identical code for reading public
and private keys from a file specified on the command line
or key contents from an environment variable, and constructing
a sumdb.note.Verifier or Signer. This change writes some common
functions that can be used for this purpose.

The motivation here, aside from just avoiding code duplication,
is that I want to add a capability for external signing and
verification tools via an external program. This could be used,
for example, to offload the actual signing to something like
an HSM or software key vault. Having the Signer/Verifier
construction in one place means there's only one place to change
to add this functionality.