Implement DSM signing with approvals for Trillian tool. HRST-28.
My previous change to the Trillian tooling added the ability
to sign with DSM. This change extends that capability to sign
using keys that require quorum approvals.

The integrate command is the only one that performs signing.
It will automatically try to sign without an approval. It will
only attempt to sign with an approval if signing without
the approval returns an error indicating that an approval
is required.

This can be used in two different modes. By default, the
integrate command will continue running and will wait util
the request is either approved or denied. Alternatively,
you can pass the --suspend option with a file name. In this
case, if approval is required, some state will be written to
the named file. After the signing request is approved, you
can resume signing with --resume, and the operation will
be completed with the signature produced by DSM.

I've also added a command line tool for approving signing
requests. This is intended to be used by the
serverless-test-dsm-approvals.sh test script with an appropriately
configured group so the requested signatures for the test
can be approved without having someone manually approve them
in the UI.