Print more information when signing log roots. HRST-64.
This adds some more output to the transparency log "integrate"
command when it is signing the new tree root. This is helpful when
signing with a quorum-protected key, as it allows the signers
to verify that the signing request matches what the signing
job in Jenkins is actually trying to sign.

After the code change merges, I will need to update the executables
in toolchain, and then make changes to the signing scripts to
use the new tool version, before we can use the new binaries
in the release signing job.