Provide an app-specific index for transparency log entries. HRST-70.
The transparency infrastructure currently only provides efficient
lookup for entries if you have either the entire entry, or at least
the node hash of the entry (which is not simply the sha256 hash
of the entry). What we want instead is the ability to look up
entries by a subset of their contents. Specifically in our case, we
would like to find the entries with a specific tuple of
(mrenclave, mrsigner, isvprodid, isvsvn), without needing to know
any of the other entry values.

To allow for this, I've added a mechanism for an app-specific
identifier to be associated with an entry when it's added, and
the ability to look up entries by their app-specific identifier.
It's up to the client adding entries and looking up entries to
compute this identifier themselves. I could have made the transparency
tooling able to compute these identifiers themselves, but the tooling
currently doesn't need to know anything about the contents of the log
entries, and I would like to keep it that way.

The index that allows looking up by app-specific identifier is
in addition to the normal transparency data structures. 3rd party
log monitoring utilities should be able to monitor our logs
and simply ignore the additional index. The new index just provides
us a mechanism to find particular entries based on a different lookup
key, and doesn't affect the security of the append-only logs in any
way.